+++ openquake/server/views.py
@@ -791,7 +791,7 @@ def aristotle_run(request):
     """
     rupture_path = get_uploaded_file_path(request, 'rupture_file')
     station_data_path = get_uploaded_file_path(request, 'station_data_file')
-    rup, rupdic, params, err = aristotle_validate(
+    _rup, rupdic, params, err = aristotle_validate(
         request.POST, rupture_path, station_data_path)
     if err:
         return HttpResponse(content=json.dumps(err), content_type=JSON,
